Prepare command
