+++ Cargo.toml
@@ -48,6 +48,7 @@ serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 tracing = "0.1.41"
 tracing-subscriber = { version = "0.3.18", features = ["env-filter", "ansi"] }
+tracing-tree = "0.4.0"
 base64 = "0.22.1"
 ruint = { version = "1.12.3", features = ["num-traits", "rand"] }
 rand = "0.8.5"
@@ -57,7 +58,6 @@ hex-literal = "0.4.1"
 itertools = "0.13.0"
 flate2 = "1.0"
 bincode = "1.3"
-
 
 # Noir lang: make sure it matches installed version `noirup -C 03b58fa2`
 acir = { git = "https://github.com/noir-lang/noir", rev = "03b58fa2" }
