+++ noir-examples/poseidon-rounds/Prover.toml
@@ -1,7 +1,7 @@
-plains = [1,2]
+plains = [1, 2]
 a = 1
 b = 2
 c = 3
 d = 5
 x = 0
-result = '0x0b5b6e4be36ffe606c2f640d972634fcf030bd0346eadcec12d0e3b95bf7f523'
+result = '0x0b50a50fda9459e4eda4b0e9382fcbb5d00c6b387bd6b79b9efb3ff95c2a32b6'
