+++ noir-examples/poseidon-rounds/src/main.nr
@@ -2,7 +2,7 @@ use dep::poseidon2;
 
 fn main(plains: [Field; 2], result: Field) {
     let mut hash = poseidon2::bn254::hash_2(plains);
-    let rounds = 1000;
+    let rounds = 100;
     for _ in 0..rounds {
         hash = poseidon2::bn254::hash_1([hash]);
     }
