+++ noir-r1cs/Cargo.toml
@@ -13,6 +13,7 @@ serde_json.workspace = true
 anyhow.workspace = true
 tracing.workspace = true
 tracing-subscriber.workspace = true
+tracing-tree.workspace = true
 base64.workspace = true
 clap.workspace = true
 zeroize.workspace = true
