+++ noir-r1cs/src/compiler.rs
@@ -270,10 +270,12 @@ impl R1CS {
     }
 }
 
-/// Print the R1CS matrices and the ACIR -> R1CS witness map, useful for debugging.
+/// Print the R1CS matrices and the ACIR -> R1CS witness map, useful for
+/// debugging.
 impl std::fmt::Display for R1CS {
     fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        writeln!(f,
+        writeln!(
+            f,
             "R1CS: {} witnesses, {} constraints",
             self.witnesses, self.constraints
         )?;
