+++ noir-r1cs/src/main.rs
@@ -5,24 +5,19 @@ mod sparse_matrix;
 mod utils;
 
 use {
-    self::{cmd::Command, compiler::R1CS, sparse_matrix::SparseMatrix},
-    acir::{native_types::Witness, AcirField, FieldElement},
-    anyhow::{ensure, Context, Result as AnyResult},
-    noirc_artifacts::program::ProgramArtifact,
-    rand::Rng,
-    std::{
-        fs::File,
-        path::{Path, PathBuf},
-        vec,
+    self::{cmd::Command, sparse_matrix::SparseMatrix},
+    anyhow::Result,
+    tracing::level_filters::LevelFilter,
+    tracing_subscriber::{
+        self, fmt::format::FmtSpan, layer::SubscriberExt as _, util::SubscriberInitExt as _,
+        EnvFilter,
     },
-    tracing::{info, level_filters::LevelFilter},
-    tracing_subscriber::{self, fmt::format::FmtSpan, EnvFilter},
-    utils::{file_io::deserialize_witness_stack, PrintAbi},
+    tracing_tree::{time::Uptime, HierarchicalLayer},
 };
 
-fn main() -> AnyResult<()> {
+fn main() -> Result<()> {
     tracing_subscriber::fmt()
-        .with_span_events(FmtSpan::ACTIVE)
+        .with_span_events(FmtSpan::NONE)
         .with_ansi(true)
         .with_env_filter(
             EnvFilter::builder()
@@ -30,145 +25,11 @@ fn main() -> AnyResult<()> {
                 .with_env_var("PROVEKIT_LOG")
                 .from_env_lossy(),
         )
+        .finish()
+        .with(HierarchicalLayer::new(2).with_timer(Uptime::default()))
         .init();
 
     // Run CLI command
     let args = argh::from_env::<cmd::Args>();
     args.run()
-}
-
-fn prove_verify(program_path: &Path, witness_path: &Path) -> AnyResult<()> {
-    info!("Loading Noir program {:?}", program_path);
-    let file = File::open(program_path).context("while opening Noir program")?;
-    let program: ProgramArtifact =
-        serde_json::from_reader(file).context("while reading Noir program")?;
-
-    info!("Program noir version: {}", program.noir_version);
-    info!("Program entry point: fn main{};", PrintAbi(&program.abi));
-    ensure!(
-        program.bytecode.functions.len() == 1,
-        "Program must have one entry point."
-    );
-    let main = &program.bytecode.functions[0];
-    let num_public_parameters = main.public_parameters.0.len();
-    let num_acir_witnesses = main.current_witness_index as usize;
-    info!(
-        "ACIR: {} witnesses, {} opcodes.",
-        num_acir_witnesses,
-        main.opcodes.len()
-    );
-
-    let mut witness_stack: acir::native_types::WitnessStack<FieldElement> =
-        deserialize_witness_stack(witness_path.to_str().unwrap())?;
-
-    let witness_stack = witness_stack.pop().unwrap().witness;
-
-    if num_acir_witnesses < 15 {
-        println!("ACIR witness values:");
-        (0..num_acir_witnesses).for_each(|i| {
-            println!("{}: {:?}", i, witness_stack[&Witness(i as u32)]);
-        });
-    }
-
-    // Create the R1CS relation
-    let mut r1cs = R1CS::new();
-    r1cs.add_circuit(main);
-    print!("{}", r1cs);
-
-    // Compute a satisfying witness
-    let mut witness = vec![None; r1cs.witnesses];
-    witness[0] = Some(FieldElement::one()); // Constant
-
-    // Fill in R1CS witness values with the pre-computed ACIR witness values
-    for (acir_witness_idx, witness_idx) in &r1cs.remap {
-        witness[*witness_idx] = Some(witness_stack[&Witness(*acir_witness_idx as u32)]);
-    }
-
-    // Solve constraints (this is how Noir expects it to be done, judging from ACVM)
-    for row in 0..r1cs.constraints {
-        let [a, b, c] =
-            [&r1cs.a, &r1cs.b, &r1cs.c].map(|mat| sparse_dot(mat.iter_row(row), &witness));
-        let (val, mat) = match (a, b, c) {
-            (Some(a), Some(b), Some(c)) => {
-                assert_eq!(a * b, c, "Constraint {row} failed");
-                continue;
-            }
-            (Some(a), Some(b), None) => (a * b, &r1cs.c),
-            (Some(a), None, Some(c)) => (c / a, &r1cs.b),
-            (None, Some(b), Some(c)) => (c / b, &r1cs.a),
-            _ => {
-                dbg!(a, b, c);
-                panic!("Can not solve constraint {row}.")
-            }
-        };
-        let Some((col, val)) = solve_dot(mat.iter_row(row), &witness, val) else {
-            panic!("Could not solve constraint {row}.")
-        };
-        witness[col] = Some(val);
-    }
-
-    // Complete witness with entropy.
-    // TODO: Use better entropy source and proper sampling.
-    let mut rng = rand::thread_rng();
-    let witness = witness
-        .iter()
-        .map(|f| f.unwrap_or_else(|| FieldElement::from(rng.gen::<u128>())))
-        .collect::<Vec<_>>();
-
-    // Verify
-    let a = mat_mul(&r1cs.a, &witness);
-    let b = mat_mul(&r1cs.b, &witness);
-    let c = mat_mul(&r1cs.c, &witness);
-    a.iter()
-        .zip(b.iter())
-        .zip(c.iter())
-        .enumerate()
-        .for_each(|(row, ((&a, &b), &c))| {
-            assert_eq!(a * b, c, "Constraint {row} failed");
-        });
-
-    r1cs.write_json_to_file(num_public_parameters, &witness, "r1cs.json")?;
-
-    Ok(())
-}
-
-// Sparse dot product. `a` is assumed zero. `b` is assumed missing.
-fn sparse_dot<'a>(
-    a: impl Iterator<Item = (usize, &'a FieldElement)>,
-    b: &[Option<FieldElement>],
-) -> Option<FieldElement> {
-    let mut accumulator = FieldElement::zero();
-    for (col, &a) in a {
-        accumulator += a * b[col]?;
-    }
-    Some(accumulator)
-}
-
-// Returns a pair (i, f) such that, setting `b[i] = f`,
-// ensures `sparse_dot(a, b) = r`.
-fn solve_dot<'a>(
-    a: impl Iterator<Item = (usize, &'a FieldElement)>,
-    b: &[Option<FieldElement>],
-    r: FieldElement,
-) -> Option<(usize, FieldElement)> {
-    let mut accumulator = -r;
-    let mut missing = None;
-    for (col, &a) in a {
-        if let Some(b) = b[col] {
-            accumulator += a * b;
-        } else if missing.is_none() {
-            missing = Some((col, a));
-        } else {
-            return None;
-        }
-    }
-    missing.map(|(col, coeff)| (col, -accumulator / coeff))
-}
-
-fn mat_mul(a: &SparseMatrix<FieldElement>, b: &[FieldElement]) -> Vec<FieldElement> {
-    let mut result = vec![FieldElement::zero(); a.rows];
-    for ((i, j), &value) in a.iter() {
-        result[i] += value * b[j];
-    }
-    result
 }
